fix: checkbox test units failed
