+++ packages/components/checkbox/src/Checkbox.vue
@@ -29,9 +29,13 @@
   </label>
 </template>
 <script lang="ts">
-import { defineComponent } from 'vue'
-import { PropTypes, withUndefined } from '@idux/cdk/utils'
-import { setup } from './checkboxSetup'
+import type { CheckboxProps, CheckValue } from './types'
+
+import { defineComponent, computed, ComputedRef, inject, Ref, ref, watch, getCurrentInstance } from 'vue'
+import { PropTypes, withUndefined, isArray, hasSlot } from '@idux/cdk/utils'
+import { useAttrs } from '@idux/components/utils/'
+import { checkboxGroupInjectionKey, subjectInjectKey } from './checkbox'
+
 export default defineComponent({
   name: 'IxCheckbox',
   inheritAttrs: false,
@@ -45,6 +49,131 @@ export default defineComponent({
     value: PropTypes.string,
   },
   emits: ['update:checked', 'change'],
-  setup,
+  setup(props: CheckboxProps, { slots, emit }) {
+    const groupSubject = inject(subjectInjectKey, null)
+    const checkeValue = ref(props.checked ?? props.falseValue)
+
+    watch(
+      () => props.checked,
+      v => {
+        checkeValue.value = v ?? props.falseValue
+      },
+    )
+
+    const handleChange = (e: Event) => {
+      const targetChecked = (e.target as HTMLInputElement).checked
+      const targetCheckValue = targetChecked ? props.trueValue : props.falseValue
+      // no value props passed
+      if (props.checked === void 0) {
+        checkeValue.value = targetCheckValue
+      }
+
+      emit('update:checked', targetCheckValue)
+      emit('change', targetCheckValue)
+      groupSubject?.dispatch((props.value ?? targetCheckValue) as string)
+    }
+
+    const hasDefaultSlot = computed(() => hasSlot(slots))
+
+    const isDisabled = useDisabled(props)
+
+    const isReadonly = useReadonly(props)
+
+    const isChecked = useChecked(checkeValue)
+
+    const classes = useClasses(isDisabled, isChecked, isReadonly)
+
+    const inputName = useName()
+
+    const attrs = useAttrs({ keys: ['type', 'tabindex'] })
+
+    const inputRef = ref(null as unknown as HTMLInputElement)
+
+    const handleClick = (e: Event) => {
+      if (isReadonly.value) {
+        e.preventDefault()
+      }
+    }
+
+    const focus = () => {
+      if (!isReadonly.value) {
+        inputRef.value.focus()
+      }
+    }
+
+    const blur = () => {
+      inputRef.value.blur()
+    }
+
+    return {
+      isChecked,
+      isDisabled,
+      isReadonly,
+      hasDefaultSlot,
+      handleChange,
+      classes,
+      inputName,
+      attrs,
+      inputRef,
+      focus,
+      blur,
+      handleClick,
+    }
+  },
 })
+
+const useChecked = (checkValue: Ref<CheckValue>) => {
+  const groupProps = inject(checkboxGroupInjectionKey, {})
+  const { props } = getCurrentInstance()!
+
+  return computed(() => {
+    const groupValue = groupProps.value
+    if (groupValue && isArray(groupValue)) {
+      return groupValue.includes((props.value ?? checkValue.value) as string)
+    }
+
+    return checkValue.value === props.trueValue
+  })
+}
+
+const useDisabled = (props: CheckboxProps) => {
+  const groupProps = inject(checkboxGroupInjectionKey, {})
+  return computed(() => {
+    return !!(groupProps.disabled || props.disabled)
+  })
+}
+
+const useReadonly = (props: CheckboxProps) => {
+  const groupProps = inject(checkboxGroupInjectionKey, {})
+  return computed(() => {
+    return !!(groupProps.readonly || props.readonly)
+  })
+}
+
+const useClasses = (
+  isDisabled: ComputedRef<boolean>,
+  isChecked: ComputedRef<boolean>,
+  isReadonly: ComputedRef<boolean>,
+) => {
+  const { props } = getCurrentInstance()!
+
+  return computed(() => {
+    const disabled = isDisabled.value
+    const checked = isChecked.value
+    const readonly = isReadonly.value
+    const indeterminate = props.indeterminate
+
+    return {
+      'ix-checkbox-disabled': !!disabled,
+      'ix-checkbox-readonly': !!readonly,
+      'ix-checkbox-indeterminate': !!indeterminate,
+      'ix-checkbox-checked': !indeterminate && checked,
+    }
+  })
+}
+
+const useName = () => {
+  const groupProps = inject(checkboxGroupInjectionKey, {})
+  return computed(() => groupProps.name as string)
+}
 </script>
